services/ai/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
starlette>=0.27.0

# Database dependencies
redis>=5.0.0
motor>=3.3.0
pymongo>=4.6.0

# HTTP and async dependencies
httpx>=0.25.0
aiofiles>=23.2.0

# Basic ML dependencies
numpy>=1.24.0
pillow>=10.1.0

# Logging
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment and configuration
python-dotenv>=1.0.0
